access checks -> campaigns

+++ campaigns/views.py
@@ -25,6 +25,9 @@ from directory.helpers import disallowChanges, userDateParse
 
 import re
 
+from robinboardAPI.permissions import check_user_access, BadAccess
+
+
 class JSONResponse(HttpResponse):
 	"""
 	An HttpResponse that renders its content into JSON.
@@ -100,8 +103,11 @@ def billboardPhoto(request):
 		data = JSONParser().parse(request)#parse incoming data
 		try:
 			billboard = models.Billboard.objects.get(id=data.get('billboard'))
+			check_user_access(request,billboard)
 		except models.Billboard.DoesNotExist:
 			return HttpResponse("No billboard at " + str(data.get('billboard')),status=404)
+		except BadAccess as e:
+			return HttpResponse(e,status=401)
 		try:
 			r = imgur.createPhotoImgur(data.get('photo'))
 			photo = {
@@ -152,8 +158,11 @@ def billboardMedia(request):
 		data = disallowChanges(disallowed,data)
 		try:
 			b = models.BillboardMedia.objects.get(id=data.get("id"))
+			check_user_access(request,b)
 		except models.BillboardMedia.DoesNotExist:
 			return HttpResponse(data,status=404)
+		except BadAccess as e:
+			return HttpResponse(e,status=401)
 		serializer = serializers.BillboardMediaSerializer(b,data=data,partial=True)
 		if serializer.is_valid():
 			serializer.save()
@@ -162,12 +171,15 @@ def billboardMedia(request):
 		data = JSONParser().parse(request)
 		try:
 			b = models.BillboardMedia.objects.get(id=data.get('id'))
+			check_user_access(request,b)
 			#delete photo on imgur
 			serializer = serializers.BillboardMediaSerializer(b)
 			b.delete()#delete billboard
 			return JSONResponse(serializer.data,status=200)
 		except models.BillboardMedia.DoesNotExist:
 			return HttpResponse("No billboard media for this id",status=404)
+		except BadAccess as e:
+			return HttpResponse(e,status=401)
 	return JSONResponse(serializer.errors, status=400)
 @api_view(['POST','GET','DELETE'])
 def photo(request):
@@ -189,6 +201,7 @@ def photo(request):
 		elif request.method == 'GET':
 				try:
 					b = models.Photo.objects.get(id=request.query_params.get('id'))
+					check_user_access(request,b)
 					serializer = serializers.PhotoSerializer(b)
 					#remove sec sensitive
 					d= serializer.data
@@ -197,10 +210,13 @@ def photo(request):
 					return JSONResponse(d, status=200)
 				except models.Photo.DoesNotExist:
 					return HttpResponse("No photo at this id.", status=404)
+				except BadAccess as e:
+					return HttpResponse(e,status=401)
 		elif request.method == 'DELETE':
 			data = JSONParser().parse(request)
 			try:
 				b = models.Photo.objects.get(id=data.get('id'))
+				check_user_access(request,b)
 				#delete photo on imgur
 				serializer = serializers.PhotoSerializer(b)
 				imgur.deletePhotoImgur(b.imgurDeleteHash)
@@ -208,6 +224,8 @@ def photo(request):
 				return JSONResponse(serializer.data,status=200)
 			except models.Photo.DoesNotExist:
 				return HttpResponse("No photo for this id",status=404)
+			except BadAccess as e:
+				return HttpResponse(e,status=401)
 		return JSONResponse(serializer.errors, status=400)
 
 @api_view(['POST','GET','PUT','DELETE'])
@@ -239,6 +257,7 @@ def billboard(request):
 				response = {}
 				try:
 					b = models.Billboard.objects.get(id=request.query_params.get('id'))
+					check_user_access(request,b)
 					serializer = serializers.BillboardSerializer(b)
 					response["billboard"] = serializer.data
 					medias = models.BillboardMedia.objects.filter(billboard=b.id)
@@ -255,14 +274,19 @@ def billboard(request):
 					return JSONResponse(response, status=200)
 				except models.Billboard.DoesNotExist:
 					return HttpResponse("No billboard with this id",status=404)
+				except BadAccess as e:
+					return HttpResponse(e,status=401)
 		elif request.method == 'PUT':
 			data = JSONParser().parse(request)#parse incoming data
 			disallowed = ["company",]
 			data = disallowChanges(disallowed,data)
 			try:
 				b = models.Billboard.objects.get(id=data.get("id"))
+				check_user_access(request,b)
 			except models.Billboard.DoesNotExist:
 				return HttpResponse(data,status=404)
+			except BadAccess as e:
+				return HttpResponse(e,status=401)
 			#TODO convert to timezone aware
 			try:
 				start = data.get('start')
@@ -288,11 +312,14 @@ def billboard(request):
 			data = JSONParser().parse(request)#parse incoming data
 			try:
 				b = models.Billboard.objects.get(id=data.get('id'))
+				check_user_access(request,b)
 				serializer = serializers.BillboardSerializer(b)
 				b.delete()
 				return JSONResponse(serializer.data,status=200)
 			except models.Billboard.DoesNotExist:
 				return HttpResponse(data,status=404)
+			except BadAccess as e:
+				return HttpResponse(e,status=401)
 		else:
 			return HttpResponse(status=404)
 		return JSONResponse(serializer.errors, status=400)
@@ -310,11 +337,15 @@ def billboard(request):
 			for billboard in data:
 				try:
 					b = models.Billboard.objects.get(id=billboard)
+					check_user_access(request,b)
 					serializer = serializers.BillboardSerializer(b)
 					b.delete()
 					response.append(serializer.data)
 				except models.Billboard.DoesNotExist:
 					continue#effectively do nothing
+				except BadAccess as e:
+					print # coding=utf-8
+					continue
 			if len(response) is 0:
 				return HttpResponse(data,status=404)
 			else:
